+++ activesupport/lib/active_support/cache/null_store.rb
@@ -12,6 +12,13 @@ module ActiveSupport
     class NullStore < Store
       prepend Strategy::LocalCache
 
+      # Advertise that this cache store can be used
+      # with "recyclable cache keys" otherwise known
+      # as cache versioning.
+      def supports_in_cache_versioning?
+        true
+      end
+
       def clear(options = nil)
       end